pyproject.toml
[project]
authors = [
  {name = "Alexander Schulze", email = "alexander.schulze@bih-charite.de"},
]
classifiers = [
  "Programming Language :: Python :: 3",
  "Operating System :: OS Independent",
]
dependencies = [
  "fastapi~=0.79.0",
  "fhir.resources~=6.4.0",
  "Jinja2~=3.1.2",
  "pycap~=2.1.0",
  "python-jose~=3.3.0",
  "PyYAML~=6.0",
  "uvicorn[standard]~=0.18.2",
]
description = "A FHIR facade for different APIs"
name = "fasadeonfhir"
readme = "README.md"
requires-python = ">=3.10"
version = "0.0.1"

[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[tool.hatch.build]
packages = ["fasadeonfhir"]

[tool.black]
line-length = 100

[tool.pytest.ini_options]
addopts = "-ra -q"
pythonpath = [
  ".",
]
testpaths = [
  "./fasadeonfhir/tests",
]
